Added refactor for GitHub API calls, utilities to parse headers, access to url information.
