+++ public/version.json
@@ -1,4 +1,4 @@
-{"version":"1.0.0-201f853",
-"git-sha":"201f8539b011f1fea9155926adaa08b361a95c01",
-"timestamp":"2020-11-08T09:13:33-0300"}
+{"version":"1.0.0-7c2b8b3",
+"git-sha":"7c2b8b307f71b2860bcc9c0bc7bc5a3210fa72fd",
+"timestamp":"2020-11-09T14:35:53-0300"}
